refactor: my-history 쿼리 구조 변경
