+++ apps/client/src/pages/my-history/hooks/use-my-history.ts
@@ -1,12 +1,12 @@
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query';
 
-import { MY_HISTORY_QUERY_OPTION } from '@shared/apis/my-history/my-history-queries';
+import { MY_RECORD_QUERY_OPTION } from '@shared/apis/my-history/my-record-queries';
+import { MY_SETLIST_QUERY_OPTION } from '@shared/apis/my-history/my-setlist-queries';
+import { MY_TIME_TABLE_QUERY_OPTION } from '@shared/apis/my-history/my-timetable-queries';
 import { SortOption } from '@shared/constants/sort-label';
 
 export const useMyTimeTablePreview = () => {
-  const { data } = useSuspenseQuery(
-    MY_HISTORY_QUERY_OPTION.TIME_TABLE.PREVIEW(),
-  );
+  const { data } = useSuspenseQuery(MY_TIME_TABLE_QUERY_OPTION.PREVIEW());
   return { data };
 };
 
@@ -15,24 +15,22 @@ export const useMyTimeTableOverView = (
   enabled: boolean,
 ) => {
   const { data } = useQuery(
-    MY_HISTORY_QUERY_OPTION.TIME_TABLE.OVERVIEW(sortBy, enabled),
+    MY_TIME_TABLE_QUERY_OPTION.OVERVIEW(sortBy, enabled),
   );
   return { data };
 };
 
 export const useMySetListPreview = () => {
-  const { data } = useSuspenseQuery(MY_HISTORY_QUERY_OPTION.SETLIST.PREVIEW());
+  const { data } = useSuspenseQuery(MY_SETLIST_QUERY_OPTION.PREVIEW());
   return { data };
 };
 
 export const useMySetListOverView = (sortBy: SortOption, enabled: boolean) => {
-  const { data } = useQuery(
-    MY_HISTORY_QUERY_OPTION.SETLIST.OVERVIEW(sortBy, enabled),
-  );
+  const { data } = useQuery(MY_SETLIST_QUERY_OPTION.OVERVIEW(sortBy, enabled));
   return { data };
 };
 
 export const useMyHistoryRecord = () => {
-  const { data } = useSuspenseQuery(MY_HISTORY_QUERY_OPTION.RECORD.ALL());
+  const { data } = useSuspenseQuery(MY_RECORD_QUERY_OPTION.ALL());
   return { data };
 };
